Implement few code review comments
