--- azure-media-spring-boot-starter-sample/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,27 +1,0 @@
-{"properties": [
-  {
-    "name": "azure.media.mediaServiceUri",
-    "type": "java.lang.String",
-    "description": "A description for 'azure.media.mediaServiceUri'"
-  },
-  {
-    "name": "azure.media.oAuthUri",
-    "type": "java.lang.String",
-    "description": "A description for 'azure.media.oAuthUri'"
-  },
-  {
-    "name": "azure.media.clientId",
-    "type": "java.lang.String",
-    "description": "A description for 'azure.media.clientId'"
-  },
-  {
-    "name": "azure.media.clientSecret",
-    "type": "java.lang.String",
-    "description": "A description for 'azure.media.clientSecret'"
-  },
-  {
-    "name": "azure.media.scope",
-    "type": "java.lang.String",
-    "description": "A description for 'azure.media.scope'"
-  }
-]}
