--- azure-media-spring-boot-starter/pom.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.microsoft.azure</groupId>
-        <artifactId>azure-spring-boot-starter-parent</artifactId>
-        <version>0.1.3</version>
-        <relativePath>../azure-spring-boot-starter-parent/pom.xml</relativePath>
-    </parent>
-
-    <artifactId>azure-media-spring-boot-starter</artifactId>
-    <packaging>jar</packaging>
-
-    <name>Azure Media Services Spring Boot Starter</name>
-    <description>Spring Boot Starter for Azure Media service</description>
-    <url>https://github.com/Microsoft/azure-spring-boot-starters</url>
-
-    <licenses>
-        <license>
-            <name>MIT</name>
-            <url>https://github.com/Microsoft/azure-spring-boot-starters/blob/master/LICENSE</url>
-            <distribution>repo</distribution>
-        </license>
-    </licenses>
-
-    <developers>
-        <developer>
-            <id>rajadilipkolli</id>
-            <name>Raja Dilip Kolli</name>
-            <email>rajadileepkolli@outlook.com</email>
-        </developer>
-    </developers>
-
-    <scm>
-        <connection>scm:git:git://github.com/Microsoft/azure-spring-boot-starters.git</connection>
-        <developerConnection>scm:git:ssh://github.com:Microsoft/azure-spring-boot-starters.git</developerConnection>
-        <url>https://github.com/Microsoft/azure-spring-boot-starters/tree/master</url>
-    </scm>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.microsoft.azure</groupId>
-            <artifactId>azure-media-spring-boot-autoconfigure</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
