+++ azure-mediaservices-spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
