+++ azure-mediaservices-spring-boot-starter-sample/src/main/resources/application.properties
@@ -1,0 +1,2 @@
+azure.mediaservices.account-key=put-your-azure-media-clientid-here
+azure.mediaservices.account-name=put-your-azure-media-secret-here
