+++ azure-spring-boot-starter-parent/pom.xml
@@ -159,8 +159,8 @@
         <module>../azure-servicebus-spring-boot-starter-sample</module>
         <module>../spring-data-azure-documentdb</module>
         <module>../spring-data-azure-documentdb-sample</module>
-        <module>../azure-media-spring-boot-autoconfigure</module>
-        <module>../azure-media-spring-boot-starter</module>
-        <module>../azure-media-spring-boot-starter-sample</module>
+        <module>../azure-mediaservices-spring-boot-autoconfigure</module>
+        <module>../azure-mediaservices-spring-boot-starter</module>
+        <module>../azure-mediaservices-spring-boot-starter-sample</module>
     </modules>
 </project>
