+++ pom.xml
@@ -52,8 +52,8 @@
         <azure.storage.autoconfigure.version>0.1.3</azure.storage.autoconfigure.version>
         <azure.servicebus.autoconfigure.version>0.1.3</azure.servicebus.autoconfigure.version>
         <azure.servicebus.starter.version>0.1.3</azure.servicebus.starter.version>
-        <azure.media.starter.version>0.1.3</azure.media.starter.version>
-        <azure.media.autoconfigure.version>0.1.3</azure.media.autoconfigure.version>
+        <azure.mediaservices.starter.version>0.1.3</azure.mediaservices.starter.version>
+        <azure.mediaservices.autoconfigure.version>0.1.3</azure.mediaservices.autoconfigure.version>
 
         <azure.documentdb.version>1.11.0</azure.documentdb.version>
         <azure.servicebus.version>1.0.0-PREVIEW-3</azure.servicebus.version>
@@ -115,13 +115,13 @@
             </dependency>
             <dependency>
                 <groupId>com.microsoft.azure</groupId>
-                <artifactId>azure-media-spring-boot-autoconfigure</artifactId>
-                <version>${azure.media.autoconfigure.version}</version>
+                <artifactId>azure-mediaservices-spring-boot-autoconfigure</artifactId>
+                <version>${azure.mediaservices.autoconfigure.version}</version>
             </dependency>
             <dependency>
                 <groupId>com.microsoft.azure</groupId>
-                <artifactId>azure-media-spring-boot-starter</artifactId>
-                <version>${azure.media.starter.version}</version>
+                <artifactId>azure-mediaservices-spring-boot-starter</artifactId>
+                <version>${azure.mediaservices.starter.version}</version>
             </dependency>
 
             <!-- Azure related -->
